+++ lib/Redesign/CreateRecipes/create_second_row.dart
@@ -117,79 +117,6 @@ class Saveingridient extends StatelessWidget {
   }
 }
 
-class Smalltextfield extends StatelessWidget {
-  final String labelText;
-  final String hintText;
-  final double bottom;
-  const Smalltextfield({
-    required this.labelText,
-    this.hintText = '',
-    this.bottom = 10,
-    //10
-    super.key,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    return Positioned(
-      // Set the position for the TextField
-      bottom: bottom, // Adjust these values as needed
-      left: 10, // Adjust these values as needed
-      right: 100,
-
-      child: TextField(
-        decoration: InputDecoration(
-          labelText: labelText,
-          labelStyle: StyleUtils.labelstylepop(),
-          hintText: hintText,
-          hintStyle: StyleUtils.hintstylepop(),
-          border: const OutlineInputBorder(),
-          enabledBorder: StyleUtils.enabledborderstyle(),
-        ),
-        style: StyleUtils.textfieldstylepop(),
-      ),
-    );
-  }
-}
-
-class Numberonlytextfiel extends StatelessWidget {
-  const Numberonlytextfiel({
-    super.key,
-    required this.isPortionSize,
-    required this.labelText,
-    required this.hintText,
-    this.bottom = 10,
-  });
-
-  final bool isPortionSize;
-  final String labelText;
-  final String hintText;
-  final double bottom;
-
-  @override
-  Widget build(BuildContext context) {
-    return Positioned(
-      bottom: bottom,
-      left: 10,
-      right: 100,
-      child: TextField(
-          keyboardType:
-              isPortionSize ? TextInputType.text : TextInputType.number,
-          inputFormatters:
-              isPortionSize ? [] : [FilteringTextInputFormatter.digitsOnly],
-          decoration: InputDecoration(
-            labelText: labelText,
-            labelStyle: StyleUtils.labelstylepop(),
-            hintText: hintText,
-            hintStyle: StyleUtils.hintstylepop(),
-            border: const OutlineInputBorder(),
-            enabledBorder: StyleUtils.enabledborderstyle(),
-          ),
-          style: StyleUtils.textfieldstylepop()),
-    );
-  }
-}
-
 //'Description:'
 //required this.title,
 //final String title;
